[Refact] utils 파일 정리
utils 파일이 일관성 없이 정리가 안되어있던 것을 정리
캐시 업데이티 관련된 쿼리문을 cacheUpdater 디렉토리를 만들어서 이동
