+++ web/src/cacheUpdater/index.js
@@ -1,0 +1,3 @@
+import { updateLikeOnDetailPost, updateLikeOnPostList } from './LikeHandler';
+
+export { updateLikeOnDetailPost, updateLikeOnPostList };
